requirements.txt
sentence-transformers
faiss-cpu
python-dotenv
pandas
langchain_huggingface 